Move Command contract to correct folder
