+++ app/Console/Commands/ComposerCommand.php
@@ -2,7 +2,7 @@
 
 namespace App\Console\Commands;
 
-use App\Console\Command;
+use App\Contracts\Command;
 use Illuminate\Support\Facades\Artisan;
 
 class ComposerCommand extends Command
