+++ app/Console/Commands/DevCommands.php
@@ -2,7 +2,7 @@
 
 namespace App\Console\Commands;
 
-use App\Console\Command;
+use App\Contracts\Command;
 use App\Models\Acars;
 use App\Models\Airline;
 use App\Models\Pirep;
