+++ app/Console/Commands/DevInstall.php
@@ -2,7 +2,7 @@
 
 namespace App\Console\Commands;
 
-use App\Console\Command;
+use App\Contracts\Command;
 use Modules\Installer\Services\ConfigService;
 
 /**
