+++ app/Console/Commands/ImportCsv.php
@@ -2,7 +2,7 @@
 
 namespace App\Console\Commands;
 
-use App\Console\Command;
+use App\Contracts\Command;
 use App\Services\ImportService;
 
 class ImportCsv extends Command
