+++ app/Console/Commands/ImportFromClassic.php
@@ -2,7 +2,7 @@
 
 namespace App\Console\Commands;
 
-use App\Console\Command;
+use App\Contracts\Command;
 
 class ImportFromClassic extends Command
 {
