+++ app/Console/Commands/NavdataImport.php
@@ -2,7 +2,7 @@
 
 namespace App\Console\Commands;
 
-use App\Console\Command;
+use App\Contracts\Command;
 use App\Models\Enums\NavaidType;
 use App\Models\Navdata;
