+++ app/Console/Commands/TestApi.php
@@ -2,7 +2,7 @@
 
 namespace App\Console\Commands;
 
-use App\Console\Command;
+use App\Contracts\Command;
 use GuzzleHttp\Client;
 
 class TestApi extends Command
