+++ app/Console/Commands/YamlImport.php
@@ -2,7 +2,7 @@
 
 namespace App\Console\Commands;
 
-use App\Console\Command;
+use App\Contracts\Command;
 use App\Services\DatabaseService;
 
 /**
