+++ app/Console/Cron/Monthly.php
@@ -2,7 +2,7 @@
 
 namespace App\Console\Cron;
 
-use App\Console\Command;
+use App\Contracts\Command;
 use App\Events\CronMonthly;
 
 /**
